backend/requirements.txt
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi @ file:///croot/certifi_1752653701669/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1754202591404/work
cryptography @ file:///croot/cryptography_1753084572505/work
dnspython==2.7.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1747203985407/work
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219583522/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pymongo==4.14.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
requests @ file:///croot/requests_1750426329888/work
urllib3 @ file:///croot/urllib3_1750775463400/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
